refactor(LX TDD Training): Split Tests
Moved function tests into their own individual files

ICTRAINING-29
